fix the 'releaselock' command to not crash the server.  introduced when jonathan did pluggable locking backends.  fortunately we never actually use the release command.  but we will....
fix the client to use persistent connections to lock servers, caching
them in the parent client objects.  so lock objects now have up
references to their client, and it asks for sockets.  when give one,
it caches that it has a lock on "10.0.0.34:34" only, not the actual
socket, and gets it back later on DESTROY to unlock it.

this will waste less ephemeral ports and syndication will catch back up.




git-svn-id: http://code.sixapart.com/svn/ddlockd/trunk@33 75bc8fc0-0210-0410-8e5e-a32055405bc5